requirements.txt
asn1crypto==0.24.0
cffi==1.11.5
Click==7.0
cryptography==2.4.2
enum34==1.1.6
Flask==1.0.2
Flask-SQLAlchemy==2.3.2
get==2018.11.19
gevent==1.3.7
greenlet==0.4.15
idna==2.7
ipaddress==1.0.22
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.15.4
Pillow==5.3.0
Pillow-PIL==0.1.dev0
post==2018.11.20
public==2018.11.20
pycparser==2.19
PyMySQL==0.9.2
query-string==2018.11.20
request==2018.11.20
six==1.11.0
SQLAlchemy==1.2.14
uuid==1.30
Werkzeug==0.14.1
